// packages/sheets-table-ui/src/views/components/type.ts
/**
 * Copyright 2023-present DreamNum Co., Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type { TableConditionTypeEnum, TableDateCompareTypeEnum, TableNumberCompareTypeEnum, TableStringCompareTypeEnum } from '@univerjs/sheets-table';

export enum ConditionSubComponentEnum {
    DatePicker = 'DatePicker',
    DateRange = 'DateRange',
    Input = 'Input',
    Inputs = 'Inputs',
    Select = 'Select',
    None = 'None',
}

export type IConditionCompareTypeEnum = TableStringCompareTypeEnum | TableDateCompareTypeEnum | TableNumberCompareTypeEnum;

export type ITableConditionTypeEnumWithoutLogic = Exclude<TableConditionTypeEnum, TableConditionTypeEnum.Logic>;

export interface IConditionInfo {
    type: ITableConditionTypeEnumWithoutLogic;
    compare: IConditionCompareTypeEnum;
    info: IConditionExpect;
}

export interface IConditionExpect {
    date?: Date;
    dateRange?: [Date, Date];
    dateSelect?: string;
    number?: number;
    numberRange?: [number?, number?];
    string?: string;
}